backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.9.3
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==3.7.1
attrs==25.1.0
beautifulsoup4==4.13.3
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.1
cymem==2.0.11
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.104.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
Mako==1.3.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.12
nltk==3.9.1
numpy==2.2.3
packaging==24.2
passlib[bcrypt]==1.7.4
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==11.1.0
preshed==3.0.9
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.9
pyasn1==0.4.8
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.27.2
Pygments==2.19.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-docx==1.0.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
regex==2024.11.6
reportlab==4.0.8
requests==2.31.0
rich==13.9.4
rsa==4.9
scikit-learn==1.3.2
scipy==1.15.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.5.1
starlette==0.27.0
thinc==8.3.4
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.0
urllib3==2.3.0
uvicorn==0.24.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
yarl==1.18.3
